+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Feature] Add `leftIcon` and `rightIcon` to **Button** when acting as a link ([#1386](https://github.com/optimizely/oui/pull/1386))
 
 ## 46.11.0 - 2020-07-23